refactor: colocate exposure hook with existing flag hook
Feedback: creating a dedicated hook file and test file for experiment-scoped
telemetry felt heavy for what will be deleted in weeks. Fold into the existing
flag file instead — same file owns all scaffolding for this flag, deletes
atomically when the experiment retires.

- Move useTrackDefaultPrivilegesExposure into useDataApiRevokeOnCreateDefault.ts
- Fold its tests into the existing test file as a second describe block
- Delete the now-redundant standalone files
- Update imports in the two consuming pages
