+++ apps/studio/hooks/misc/useDataApiRevokeOnCreateDefault.ts
@@ -1,5 +1,8 @@
+import { useEffect, useRef } from 'react'
+
 import { usePHFlag } from '../ui/useFlag'
 import { IS_TEST_ENV } from '@/lib/constants'
+import { useTrack } from '@/lib/telemetry/track'
 
 /**
  * Controls the default state of the "Default privileges for new entities"
@@ -21,4 +24,37 @@ export const useDataApiRevokeOnCreateDefaultEnabled = (): boolean => {
   }
 
   return !!flag
+}
+
+type DefaultPrivilegesExposureOptions =
+  | { surface: 'main'; dataApiEnabled: boolean }
+  | { surface: 'vercel' }
+
+/**
+ * Fires `project_creation_default_privileges_exposed` once per mount after the
+ * `dataApiRevokeOnCreateDefault` flag resolves. Gating on flag resolution keeps
+ * cohort attribution clean — users whose flag never resolves are not counted in
+ * either cohort. Deduplicated via ref so re-renders and mid-session flag flips
+ * don't re-fire.
+ */
+export const useTrackDefaultPrivilegesExposure = (
+  options: DefaultPrivilegesExposureOptions
+) => {
+  const track = useTrack()
+  const flag = usePHFlag<boolean>('dataApiRevokeOnCreateDefault')
+  const hasTracked = useRef(false)
+
+  const { surface } = options
+  const dataApiEnabled = options.surface === 'main' ? options.dataApiEnabled : undefined
+
+  useEffect(() => {
+    if (hasTracked.current) return
+    if (flag === undefined) return
+    hasTracked.current = true
+    track('project_creation_default_privileges_exposed', {
+      surface,
+      ...(surface === 'main' && { dataApiEnabled: dataApiEnabled as boolean }),
+      dataApiRevokeOnCreateDefaultEnabled: flag,
+    })
+  }, [flag, track, surface, dataApiEnabled])
 }
